pyproject.toml
[project]
name = "snapforge-api"
version = "0.1.0"
description = "File processing micro-SaaS API — screenshots, QR codes, OG images, image processing, PDF generation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pillow>=10.0",
    "qrcode[pil]>=7.4",
    "python-multipart>=0.0.9",
    "playwright>=1.40",
    "jinja2>=3.1",
    "weasyprint>=62.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio"]

[tool.ruff]
line-length = 120
target-version = "py311"
